+++ runtime/t2rn-parachain/src/circuit_config.rs
@@ -71,7 +71,9 @@ impl pallet_attesters::Config for Runtime {
     type Xdns = XDNS;
 }
 
-use t3rn_primitives::{monetary::UNIT as TRN, xdns::PalletAssetsOverlay};
+use t3rn_primitives::xdns::PalletAssetsOverlay;
+use circuit_runtime_types::UNIT as TRN;
+
 
 parameter_types! {
     pub const TotalInflation: Perbill = Perbill::from_parts(44_000_000); // 4.4%