+++ runtime/t2rn-parachain/src/lib.rs
@@ -11,15 +11,15 @@ include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 pub const VERSION: RuntimeVersion = RuntimeVersion {
     spec_name: create_runtime_str!("t2rn"),
     impl_name: create_runtime_str!("t2rn"),
-    authoring_version: 37,
-    spec_version: 37,
-    impl_version: 37,
+    authoring_version: 39,
+    spec_version: 39,
+    impl_version: 39,
     apis: RUNTIME_API_VERSIONS,
-    transaction_version: 37,
+    transaction_version: 39,
     state_version: 1,
 };
 
-use t3rn_primitives::monetary::MILLIUNIT as BASE_MILLIUNIT;
+use circuit_runtime_types::MILLIUNIT as BASE_MILLIUNIT;
 
 use frame_system::EnsureRoot;