+++ runtime/t7rn-parachain/src/lib.rs
@@ -50,11 +50,11 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
     // https://docs.rs/sp-version/latest/sp_version/struct.RuntimeVersion.html
     spec_name: create_runtime_str!("t1rn"),
     impl_name: create_runtime_str!("t1rn Circuit Collator"),
-    authoring_version: 15,
-    spec_version: 15,
-    impl_version: 15,
+    authoring_version: 17,
+    spec_version: 17,
+    impl_version: 17,
     apis: RUNTIME_API_VERSIONS,
-    transaction_version: 16,
+    transaction_version: 18,
     // https://github.com/paritytech/cumulus/issues/998
     // https://github.com/paritytech/substrate/pull/9732
     // https://github.com/paritytech/substrate/pull/10073