+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.88.1-rc.0] - 2024-05-02
+### :bug: Bug Fixes
+- [`f6ed048`](https://github.com/t3rn/t3rn/commit/f6ed048df01faaf6083c3d77f972beb33dabaaa9) - guardian integration flow gas price error *(PR [#1600](https://github.com/t3rn/t3rn/pull/1600) by [@chexware](https://github.com/chexware))*
+
+
+## [v1.88.0-rc.0] - 2024-04-17
+### :sparkles: New Features
+- [`512e64e`](https://github.com/t3rn/t3rn/commit/512e64e8edb44b20684c29f0ceb10dc6cd37173a) - added GRANDPA RPC config to t2rn *(PR [#1599](https://github.com/t3rn/t3rn/pull/1599) by [@chexware](https://github.com/chexware))*
+
+
 ## [v1.87.4-rc.0] - 2024-04-12
 ### :bug: Bug Fixes
 - [`3388c47`](https://github.com/t3rn/t3rn/commit/3388c470d6589edc8fc64e5ab6e0cac6af08f712) - updated native token from 12 to 18 decimals *(PR [#1597](https://github.com/t3rn/t3rn/pull/1597) by [@chexware](https://github.com/chexware))*
@@ -2097,3 +2107,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v1.87.2-rc.0]: https://github.com/t3rn/t3rn/compare/v1.87.1-rc.0...v1.87.2-rc.0
 [v1.87.3-rc.0]: https://github.com/t3rn/t3rn/compare/v1.87.2-rc.0...v1.87.3-rc.0
 [v1.87.4-rc.0]: https://github.com/t3rn/t3rn/compare/v1.87.3-rc.0...v1.87.4-rc.0
+[v1.88.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.87.4-rc.0...v1.88.0-rc.0
+[v1.88.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.88.0-rc.0...v1.88.1-rc.0